Backend/requirements.txt
fastapi
uvicorn
google-generativeai
requests
statsmodels
numpy
pandas
python-dotenv
geocoder
joblib
scikit-learn
